requirements.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
PyJWT==2.8.0
jwt==1.3.1
psycopg2==2.9.9
